Rename EcDSA.ES256K -> Secp256k1 & EDDSA.Ed25519 -> Ed25519
