--- Sources/Web5/Crypto/DigitalSignatureAlgorithms/ECDSA.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-/// Elliptic Curve Digital Signature Algorithm (ECDSA)
-public enum ECDSA {}
