--- Sources/Web5/Crypto/DigitalSignatureAlgorithms/EdDSA.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-/// Edwards-curve Digital Signature Algorithm (EdDSA)
-public enum EdDSA {}
