+++ Tests/Web5Tests/Crypto/Algorithms/Secp256k1Tests.swift
@@ -3,12 +3,12 @@ import XCTest
 
 @testable import Web5
 
-final class ECDSA_Es256kTests: XCTestCase {
+final class Secp256k1Tests: XCTestCase {
 
     let payload = "Hello, World!".data(using: .utf8)!
 
     func test_generatePrivateKey() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
+        let privateKey = try Secp256k1.generatePrivateKey()
 
         XCTAssertEqual(privateKey.curve, .secp256k1)
         XCTAssertEqual(privateKey.keyType, .elliptic)
@@ -19,8 +19,8 @@ final class ECDSA_Es256kTests: XCTestCase {
     }
 
     func test_computePublicKey() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
-        let publicKey = try ECDSA.Es256k.computePublicKey(privateKey: privateKey)
+        let privateKey = try Secp256k1.generatePrivateKey()
+        let publicKey = try Secp256k1.computePublicKey(privateKey: privateKey)
 
         XCTAssertEqual(publicKey.curve, .secp256k1)
         XCTAssertEqual(publicKey.keyType, .elliptic)
@@ -37,48 +37,48 @@ final class ECDSA_Es256kTests: XCTestCase {
     }
 
     func test_privateKey_toAndFromBytes() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
-        let privateKeyBytes = try ECDSA.Es256k.privateKeyToBytes(privateKey)
-        let restoredPrivateKey = try ECDSA.Es256k.privateKeyFromBytes(privateKeyBytes)
+        let privateKey = try Secp256k1.generatePrivateKey()
+        let privateKeyBytes = try Secp256k1.privateKeyToBytes(privateKey)
+        let restoredPrivateKey = try Secp256k1.privateKeyFromBytes(privateKeyBytes)
 
         XCTAssertNoDifference(privateKey, restoredPrivateKey)
     }
 
     func test_publicKey_toAndFromBytes() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
-        let publicKey = try ECDSA.Es256k.computePublicKey(privateKey: privateKey)
-        let publicKeyBytes = try ECDSA.Es256k.publicKeyToBytes(publicKey)
-        let restoredPublicKey = try ECDSA.Es256k.publicKeyFromBytes(publicKeyBytes)
+        let privateKey = try Secp256k1.generatePrivateKey()
+        let publicKey = try Secp256k1.computePublicKey(privateKey: privateKey)
+        let publicKeyBytes = try Secp256k1.publicKeyToBytes(publicKey)
+        let restoredPublicKey = try Secp256k1.publicKeyFromBytes(publicKeyBytes)
 
         XCTAssertNoDifference(publicKey, restoredPublicKey)
     }
 
     func test_sign() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
-        let signature = try ECDSA.Es256k.sign(payload: payload, privateKey: privateKey)
+        let privateKey = try Secp256k1.generatePrivateKey()
+        let signature = try Secp256k1.sign(payload: payload, privateKey: privateKey)
 
         /// Signatures should always be 64 bytes in length
         XCTAssertEqual(signature.count, 64)
     }
 
     func test_sign_errorsWhenPrivateKeyFromAnotherAlgorithm() throws {
-        let ed25519PrivateKey = try EdDSA.Ed25519.generatePrivateKey()
-        XCTAssertThrowsError(try ECDSA.Es256k.sign(payload: payload, privateKey: ed25519PrivateKey))
+        let ed25519PrivateKey = try Ed25519.generatePrivateKey()
+        XCTAssertThrowsError(try Secp256k1.sign(payload: payload, privateKey: ed25519PrivateKey))
     }
 
     func test_verify() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
-        let publickey = try ECDSA.Es256k.computePublicKey(privateKey: privateKey)
-        let signature = try ECDSA.Es256k.sign(payload: payload, privateKey: privateKey)
-        let isValid = try ECDSA.Es256k.verify(payload: payload, signature: signature, publicKey: publickey)
+        let privateKey = try Secp256k1.generatePrivateKey()
+        let publickey = try Secp256k1.computePublicKey(privateKey: privateKey)
+        let signature = try Secp256k1.sign(payload: payload, privateKey: privateKey)
+        let isValid = try Secp256k1.verify(payload: payload, signature: signature, publicKey: publickey)
 
         XCTAssertTrue(isValid)
     }
 
     func test_verify_invalidWhenPayloadMutated() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
-        let publickey = try ECDSA.Es256k.computePublicKey(privateKey: privateKey)
-        let signature = try ECDSA.Es256k.sign(payload: payload, privateKey: privateKey)
+        let privateKey = try Secp256k1.generatePrivateKey()
+        let publickey = try Secp256k1.computePublicKey(privateKey: privateKey)
+        let signature = try Secp256k1.sign(payload: payload, privateKey: privateKey)
 
         // Make a copy and flip the least significant bit of the payload
         var mutatedPayload = payload
@@ -86,21 +86,21 @@ final class ECDSA_Es256kTests: XCTestCase {
 
         // Verification should return false, as the verified payload does not
         // match the payload used to generate signature
-        let isValid = try ECDSA.Es256k.verify(payload: mutatedPayload, signature: signature, publicKey: publickey)
+        let isValid = try Secp256k1.verify(payload: mutatedPayload, signature: signature, publicKey: publickey)
         XCTAssertFalse(isValid)
     }
 
     func test_verify_invalidWhenSignatureMutated() throws {
-        let privateKey = try ECDSA.Es256k.generatePrivateKey()
-        let publickey = try ECDSA.Es256k.computePublicKey(privateKey: privateKey)
-        let signature = try ECDSA.Es256k.sign(payload: payload, privateKey: privateKey)
+        let privateKey = try Secp256k1.generatePrivateKey()
+        let publickey = try Secp256k1.computePublicKey(privateKey: privateKey)
+        let signature = try Secp256k1.sign(payload: payload, privateKey: privateKey)
 
         // Make a copy and flip the least significant bit of the signature
         var mutatedSignature = Data(signature)
         mutatedSignature[0] ^= 1 << 0
 
         // Verification should return false, as the signature for the payload has been mutated
-        let isValid = try ECDSA.Es256k.verify(payload: payload, signature: mutatedSignature, publicKey: publickey)
+        let isValid = try Secp256k1.verify(payload: payload, signature: mutatedSignature, publicKey: publickey)
         XCTAssertFalse(isValid)
     }
 }
